+++ src/test/scala/au/com/agl/arc/extract/AvroExtractSuite.scala
@@ -20,6 +20,7 @@ class AvroExtractSuite extends FunSuite with BeforeAndAfter {
 
   var session: SparkSession = _  
   val targetFile = FileUtils.getTempDirectoryPath() + "extract.avro" 
+  val targetFileGlob = FileUtils.getTempDirectoryPath() + "ex{t,a,b,c}ract.avro" 
   val emptyDirectory = FileUtils.getTempDirectoryPath() + "empty.avro" 
   val emptyWildcardDirectory = FileUtils.getTempDirectoryPath() + "*.avro.gz" 
   val outputView = "dataset"
@@ -65,7 +66,7 @@ class AvroExtractSuite extends FunSuite with BeforeAndAfter {
         name=outputView,
         cols=cols.right.getOrElse(Nil),
         outputView=outputView,
-        input=new URI(targetFile),
+        input=targetFileGlob,
         authentication=None,
         params=Map.empty,
         persist=false,
@@ -105,7 +106,7 @@ class AvroExtractSuite extends FunSuite with BeforeAndAfter {
         name=outputView,
         cols=Nil,
         outputView=outputView,
-        input=new URI(targetFile),
+        input=targetFile,
         authentication=None,
         params=Map.empty,
         persist=false,
@@ -122,7 +123,7 @@ class AvroExtractSuite extends FunSuite with BeforeAndAfter {
         name=outputView,
         cols=Nil,
         outputView=outputView,
-        input=new URI(targetFile),
+        input=targetFile,
         authentication=None,
         params=Map.empty,
         persist=true,
@@ -160,7 +161,7 @@ class AvroExtractSuite extends FunSuite with BeforeAndAfter {
           name=outputView,
           cols=Nil,
           outputView=outputView,
-          input=new URI(emptyWildcardDirectory),
+          input=emptyWildcardDirectory,
           authentication=None,
           params=Map.empty,
           persist=false,
@@ -179,7 +180,7 @@ class AvroExtractSuite extends FunSuite with BeforeAndAfter {
           name=outputView,
           cols=Nil,
           outputView=outputView,
-          input=new URI(emptyDirectory),
+          input=emptyDirectory,
           authentication=None,
           params=Map.empty,
           persist=false,
@@ -197,7 +198,7 @@ class AvroExtractSuite extends FunSuite with BeforeAndAfter {
         name=outputView,
         cols=cols,
         outputView=outputView,
-        input=new URI(emptyDirectory),
+        input=emptyDirectory,
         authentication=None,
         params=Map.empty,
         persist=false,