+++ src/test/scala/au/com/agl/arc/extract/ORCExtractSuite.scala
@@ -21,6 +21,7 @@ class ORCExtractSuite extends FunSuite with BeforeAndAfter {
 
   var session: SparkSession = _  
   val targetFile = FileUtils.getTempDirectoryPath() + "extract.orc" 
+  val targetFileGlob = FileUtils.getTempDirectoryPath() + "ex{t,a,b,c}ract.orc" 
   val emptyDirectory = FileUtils.getTempDirectoryPath() + "empty.orc" 
   val emptyWildcardDirectory = FileUtils.getTempDirectoryPath() + "*.orc.gz" 
   val outputView = "dataset"
@@ -66,7 +67,7 @@ class ORCExtractSuite extends FunSuite with BeforeAndAfter {
         name=outputView,
         cols=cols.right.getOrElse(Nil),
         outputView=outputView,
-        input=new URI(targetFile),
+        input=targetFileGlob,
         authentication=None,
         params=Map.empty,
         persist=false,
@@ -101,7 +102,7 @@ class ORCExtractSuite extends FunSuite with BeforeAndAfter {
         name=outputView,
         cols=Nil,
         outputView=outputView,
-        input=new URI(targetFile),
+        input=targetFile,
         authentication=None,
         params=Map.empty,
         persist=false,
@@ -119,7 +120,7 @@ class ORCExtractSuite extends FunSuite with BeforeAndAfter {
         name=outputView,
         cols=Nil,
         outputView=outputView,
-        input=new URI(targetFile),
+        input=targetFile,
         authentication=None,
         params=Map.empty,
         persist=true,
@@ -157,7 +158,7 @@ class ORCExtractSuite extends FunSuite with BeforeAndAfter {
           name=outputView,
           cols=Nil,
           outputView=outputView,
-          input=new URI(emptyWildcardDirectory),
+          input=emptyWildcardDirectory,
           authentication=None,
           params=Map.empty,
           persist=false,
@@ -176,7 +177,7 @@ class ORCExtractSuite extends FunSuite with BeforeAndAfter {
           name=outputView,
           cols=Nil,
           outputView=outputView,
-          input=new URI(emptyDirectory),
+          input=emptyDirectory,
           authentication=None,
           params=Map.empty,
           persist=false,
@@ -194,7 +195,7 @@ class ORCExtractSuite extends FunSuite with BeforeAndAfter {
         name=outputView,
         cols=cols,
         outputView=outputView,
-        input=new URI(emptyDirectory),
+        input=emptyDirectory,
         authentication=None,
         params=Map.empty,
         persist=false,